remove internal LPMS depenency
