--- lpms/README.md
@@ -1,14 +1,0 @@
-# Livepeer Media Server (LPMS)
-
-The Livepeer Media Server is meant to be a standalone media server
-which provides:
-
-1. Input/Output interface for video in various formats.
-2. Transcoding, Transmuxing, Transrating functionality to convert
-input video into various output encodings, formats, and bitrates.
-
-At the moment, it isn't modular enough to fully decouple from the
-go-livepeer packages, but this is a near term goal in the roadmap.
-
-
-
